+++ tests/phpunit/integration/CRUD/Rdbms/AbstractRdbmsCRUDTest.php
@@ -3,7 +3,9 @@
 namespace MediaWiki\WikispeechSpeechDataCollector\Tests\Integration\CRUD\Rdbms;
 
 use MediaWiki\MediaWikiServices;
+use MediaWiki\WikispeechSpeechDataCollector\CRUD\AbstractCRUD;
 use MediaWiki\WikispeechSpeechDataCollector\CRUD\CRUDContext;
+use MediaWiki\WikispeechSpeechDataCollector\CRUD\Rdbms\AbstractRdbmsCRUD;
 use MediaWiki\WikispeechSpeechDataCollector\Tests\Integration\CRUD\AbstractCRUDTest;
 use Wikimedia\TestingAccessWrapper;
 
@@ -14,8 +16,6 @@ use Wikimedia\TestingAccessWrapper;
  *
  * @since 0.1.0
  * @group Database
- * @covers \MediaWiki\WikispeechSpeechDataCollector\CRUD\AbstractCRUD
- * @covers \MediaWiki\WikispeechSpeechDataCollector\CRUD\CLUD
  */
 abstract class AbstractRdbmsCRUDTest extends AbstractCRUDTest {
 
@@ -26,10 +26,18 @@ abstract class AbstractRdbmsCRUDTest extends AbstractCRUDTest {
 		return $this->context;
 	}
 
+	/**
+	 * @param CRUDContext $context
+	 * @return AbstractRdbmsCRUD
+	 */
+	abstract protected function newCRUDInstance( CRUDContext $context ): AbstractCRUD;
+
 	protected function setUp(): void {
 		parent::setUp();
 		$this->context = new CRUDContext(
-			MediaWikiServices::getInstance()->getDBLoadBalancer()
+			MediaWikiServices::getInstance()->getDBLoadBalancer(),
+			null,
+			null
 		);
 		$crudWrapper = TestingAccessWrapper::newFromObject( $this->newCRUDInstance( $this->context ) );
 		$this->tablesUsed[] = $crudWrapper->getTable();